pyproject.toml
[project]
name = "notes"
version = "0.1.0"
description = "project for notes"
authors = [
    {name = "Agbar102",email = "kamchiagbar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "idna (>=3.10,<4.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
